requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
mysqlclient>=2.1,<3.0
django-environ>=0.8,<1.0
gunicorn>=20.1,<21.0
django-crispy-forms>=1.12,<2.0
django-allauth>=0.47,<1.0
celery>=5.2,<6.0
redis>=4.0,<5.0
python-dotenv>=0.19,<1.0
Pillow>=8.0,<9.0
